Move icons, refine styles, add changeset
