+++ .changeset/flat-cows-sleep.md
@@ -1,0 +1,6 @@
+---
+"nextjs-website": patch
+"strapi-cms": patch
+---
+
+Rework of StandardHeader component
