+++ apps/strapi-cms/viewconfig.json
@@ -239,7 +239,7 @@
   },
   {
     "key": "plugin_content_manager_configuration_components::menu.menu-link",
-    "value": "{\"uid\":\"menu.menu-link\",\"settings\":{\"bulkable\":true,\"filterable\":true,\"searchable\":true,\"pageSize\":10,\"mainField\":\"label\",\"defaultSortBy\":\"label\",\"defaultSortOrder\":\"ASC\"},\"metadatas\":{\"id\":{\"edit\":{},\"list\":{\"label\":\"id\",\"searchable\":false,\"sortable\":false}},\"label\":{\"edit\":{\"label\":\"Testo\",\"description\":\"\",\"placeholder\":\"\",\"visible\":true,\"editable\":true},\"list\":{\"label\":\"label\",\"searchable\":true,\"sortable\":true}},\"page\":{\"edit\":{\"label\":\"Pagina\",\"description\":\"\",\"placeholder\":\"\",\"visible\":true,\"editable\":true,\"mainField\":\"name\"},\"list\":{\"label\":\"page\",\"searchable\":true,\"sortable\":true}},\"sublinks\":{\"edit\":{\"label\":\"Sottolink\",\"description\":\"\",\"placeholder\":\"\",\"visible\":true,\"editable\":true},\"list\":{\"label\":\"sublinks\",\"searchable\":false,\"sortable\":false}},\"sectionID\":{\"edit\":{\"label\":\"ID Sezione\",\"description\":\"\",\"placeholder\":\"\",\"visible\":true,\"editable\":true},\"list\":{\"label\":\"sectionID\",\"searchable\":true,\"sortable\":true}},\"alignRight\":{\"edit\":{\"label\":\"Allinea a Destra\",\"description\":\"\",\"placeholder\":\"\",\"visible\":true,\"editable\":true},\"list\":{\"label\":\"alignRight\",\"searchable\":true,\"sortable\":true}},\"documentId\":{\"edit\":{},\"list\":{\"label\":\"documentId\",\"searchable\":true,\"sortable\":true}}},\"layouts\":{\"list\":[\"id\",\"label\",\"page\",\"sublinks\"],\"edit\":[[{\"name\":\"label\",\"size\":6},{\"name\":\"alignRight\",\"size\":6}],[{\"name\":\"page\",\"size\":6},{\"name\":\"sectionID\",\"size\":6}],[{\"name\":\"sublinks\",\"size\":12}]]},\"isComponent\":true}",
+    "value": "{\"uid\":\"menu.menu-link\",\"settings\":{\"bulkable\":true,\"filterable\":true,\"searchable\":true,\"pageSize\":10,\"mainField\":\"label\",\"defaultSortBy\":\"label\",\"defaultSortOrder\":\"ASC\"},\"metadatas\":{\"id\":{\"edit\":{},\"list\":{\"label\":\"id\",\"searchable\":false,\"sortable\":false}},\"label\":{\"edit\":{\"label\":\"Testo\",\"description\":\"\",\"placeholder\":\"\",\"visible\":true,\"editable\":true},\"list\":{\"label\":\"label\",\"searchable\":true,\"sortable\":true}},\"page\":{\"edit\":{\"label\":\"Pagina\",\"description\":\"\",\"placeholder\":\"\",\"visible\":true,\"editable\":true,\"mainField\":\"name\"},\"list\":{\"label\":\"page\",\"searchable\":true,\"sortable\":true}},\"sublinks\":{\"edit\":{\"label\":\"Sottolink\",\"description\":\"\",\"placeholder\":\"\",\"visible\":true,\"editable\":true},\"list\":{\"label\":\"sublinks\",\"searchable\":false,\"sortable\":false}},\"sectionID\":{\"edit\":{\"label\":\"ID Sezione\",\"description\":\"\",\"placeholder\":\"\",\"visible\":true,\"editable\":true},\"list\":{\"label\":\"sectionID\",\"searchable\":true,\"sortable\":true}},\"documentId\":{\"edit\":{},\"list\":{\"label\":\"documentId\",\"searchable\":true,\"sortable\":true}}},\"layouts\":{\"list\":[\"id\",\"label\",\"page\",\"sublinks\"],\"edit\":[[{\"name\":\"label\",\"size\":6}],[{\"name\":\"page\",\"size\":6},{\"name\":\"sectionID\",\"size\":6}],[{\"name\":\"sublinks\",\"size\":12}]]},\"isComponent\":true}",
     "type": "object",
     "environment": null,
     "tag": null
